misc: reduce dependabot noise
Because of the weird^H^H^H^H^Hinteresting structure of the contrib repo,
we have an overwhelming amount of dependabot noise for this. Especially
since we're behind upstream.

We could turn this off entirely, I suppose. Or - we can modify the
Makefile and just make this noise monthly rather than weekly.

This commit is the result of modifying the `Makefile` to do monthly
updates instead of weekly, as well as the result of running `make gendependabot`.